pyproject.toml
[project]
name = "fortios-mcp-server"
version = "1.0.0"
description = "MCP Server for Fortinet FortiOS 7.6.6 REST API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.9.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
fortios-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
]
